requirements.txt
django
PyPDF2
python-docx
torch
transformers
pandas
spacy
sentence_transformers
nltk